+++ src/suggest.test.ts
@@ -1,8 +1,8 @@
 import type {ApiSuggest, ApiSuggestions} from './token';
 
-export const linkSuggest: ApiSuggest<[string, number, string?]> = (s, _, ns = 0) => {
+export const linkSuggest: ApiSuggest<[string, number, (string | [string])?]> = (s, _, ns = 0) => {
 	if (ns === 0) {
-		return [[`${s} (article)`, 0], ['Alice (user)', 0, `${s} (user)`]];
+		return [[`${s} (article)`, 0], ['Alice (user)', 0, `${s} (user)`], [`${s} (disambiguation)`, 0, [`Help:${s}`]]];
 	}
 	const colon = s.indexOf(':');
 	return [[colon === -1 ? s : `${s.slice(colon + 1)} (${s.slice(0, colon).toLowerCase()})`, ns]];